feat(admin): Get postgres and sqlite both working in the admin app

+++ admin/src/performance_model.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct PerformancePageResult {
-    pub bytes: i64,
+    pub bytes: f64,
     pub url: String,
     pub display_bytes: String,
 }
@@ -10,7 +10,7 @@ pub struct PerformancePageResult {
 impl PerformancePageResult {
     pub fn default() -> PerformancePageResult {
         PerformancePageResult {
-            bytes: 0,
+            bytes: 0f64,
             url: String::from(""),
             display_bytes: String::from("N/A")
         }
