+++ admin/src/utils.rs
@@ -16,18 +16,18 @@ pub fn format_time_span(time: f64) -> String {
     format!("{}:{}:{}", hours, minutes, seconds)
 }
 
-pub fn bytes_to_display(bytes: i64) -> String {
-    if bytes < 1000i64 {
+pub fn bytes_to_display(bytes: f64) -> String {
+    if bytes < 1000f64 {
         return bytes.to_string() + "B";
     }
 
-    if bytes < 1000000i64 {
+    if bytes < 1000000f64 {
         return (bytes as f64 / 1000f64).to_string() + "Kb";
     }
 
-    if bytes < 1000000000i64 {
+    if bytes < 1000000000f64 {
         return (bytes as f64 / 1000000f64).to_string() + "Mb";
     }
 
-    (bytes / 1000000000i64).to_string() + "Gb"
+    (bytes / 1000000000f64).to_string() + "Gb"
 }
