+++ db/postgres/migrations/add_jobs.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS jobs(
     id TEXT PRIMARY KEY,
     status TEXT,
-    startTime NUMERIC,
-    totalTime NUMERIC
+    startTime DOUBLE PRECISION,
+    totalTime DOUBLE PRECISION
 );
